+++ src/KsWare.Presentation.Tests/ViewModelFramework/ViewModel/(OtherVMs)/ActionVM.Tests.cs
@@ -55,18 +55,23 @@ namespace KsWare.Presentation.Tests.ViewModelFramework {
 			x.TestPAction.Execute("A"); Assert.That(x.Result, Is.EqualTo("SampleBaseVM.DoTestP-A"));
 			x.Test3Action.Execute(); Assert.That(x.Result, Is.EqualTo("SampleVM.DoTest3"));
 			x.Test4Action.Execute(); Assert.That(x.Result, Is.EqualTo("SampleVM.DoTest4"));
+			x.Test5Action.Execute(); Assert.That(x.Result, Is.EqualTo("SampleVM.Test5"));
 		}
 
 		private class SampleVM : SampleBaseVM{
 			/// <inheritdoc />
 			public SampleVM() {
+				Test5Action = new ActionVM { MːDoAction = Test5, Parent = this }; // parent must be set after MːDoAction, because setting Parent triggers SearchAndRegisterActionMethod
 				RegisterChildren(()=>this);
 			}
 
 			public ActionVM Test3Action { get; private set; }
+			public ActionVM Test5Action { get; private set; }
+
 			private void DoTest3(){ Result = "SampleVM.DoTest3";}
 
 			protected override void DoTest4(){ Result = "SampleVM.DoTest4";}
+			private void Test5() { Result = "SampleVM.Test5"; }
 		}
 
 		private class SampleBaseVM : ObjectVM {